sizing

+++ src/util.rs
@@ -11,10 +11,10 @@ pub fn load_image(filepath: &str) -> ImageResult<DynamicImage> {
     image::open(filepath)
 }
 
-pub fn scaled_to_area(img: DynamicImage, area: u32) -> DynamicImage {
+pub fn scaled_to_area(img: DynamicImage, area: usize) -> DynamicImage {
     let w = img.width();
     let h = img.height();
-    if w * h < area {
+    if w * h < area as u32 {
         img
     } else {
         let x = ((area as f64) / (w * h) as f64).sqrt();
@@ -29,3 +29,4 @@ pub fn rng_normal(rng: &mut ThreadRng) -> f64 {
 pub fn degrees(radians: f64) -> f64 {
     radians * 180.0 / PI
 }
+
